# src/main/api/openapi.yaml
openapi: 3.0.3
info:
  title: Lyra 企业级文档管理系统 API
  description: |
    Lyra是一个企业级云原生文档管理系统，提供带有版本控制的文档管理服务。
    
    ## 核心功能
    - 用户认证与授权管理
    - 文件上传、下载、管理
    - 文件夹结构化管理
    - 文件搜索与版本控制
    - WebDAV协议支持
    - 权限管理与空间隔离
    
    ## 技术架构
    - **后端**: Spring Boot 3.5.x + Spring Security + JPA
    - **数据库**: SQLite/MySQL/PostgreSQL + Redis
    - **部署**: Docker + Kubernetes + Helm
    
    ## API特性
    - RESTful设计规范
    - JWT令牌认证
    - 统一响应格式
    - 完整的错误处理
    - 分页和排序支持
    - 文件流传输优化
    
  version: "1.0.0"
  termsOfService: https://github.com/your-org/lyra/blob/main/TERMS.md
  contact:
    name: Lyra API Support
    url: https://github.com/your-org/lyra
    email: support@lyra.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: 本地开发环境
  - url: https://api-dev.lyra.example.com
    description: 开发环境
  - url: https://api-staging.lyra.example.com  
    description: 预发布环境
  - url: https://api.lyra.example.com
    description: 生产环境

# ==================== 安全方案 ====================
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT访问令牌认证。
        
        **获取令牌**: 通过 `/api/auth/login` 端点登录获取访问令牌。
        
        **使用方式**: 在请求头中添加 `Authorization: Bearer <token>`
        
        **令牌刷新**: 使用 `/api/auth/refresh` 端点刷新过期令牌。

  # ==================== 通用响应格式 ====================
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 操作是否成功
          example: true
        message:
          type: string
          description: 响应消息
          example: "操作成功"
        data:
          description: 响应数据，根据具体接口而定
        timestamp:
          type: integer
          format: int64
          description: 时间戳(毫秒)
          example: 1642780800000
        errors:
          type: array
          description: 错误信息列表(仅在失败时出现)
          items:
            type: string
          example: ["参数验证失败", "用户名不能为空"]
      required:
        - success
        - message

    # 分页响应格式
    PageResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                content:
                  type: array
                  description: 当前页数据
                  items: {}
                totalElements:
                  type: integer
                  description: 总记录数
                  example: 100
                totalPages:
                  type: integer
                  description: 总页数
                  example: 10
                currentPage:
                  type: integer
                  description: 当前页码(从0开始)
                  example: 0
                pageSize:
                  type: integer
                  description: 每页大小
                  example: 10
                hasNext:
                  type: boolean
                  description: 是否有下一页
                  example: true
                hasPrevious:
                  type: boolean
                  description: 是否有上一页
                  example: false

    # ==================== 认证相关数据模型 ====================
    
    # 登录请求
...
    FolderInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 文件夹ID
          example: 1
        name:
          type: string
          description: 文件夹名称
          example: "重要文档"
        path:
          type: string
          description: 文件夹路径
          example: "/documents/important"
        parentId:
          type: integer
          format: int64
          description: 父文件夹ID
          nullable: true
          example: null
        spaceId:
          type: integer
          format: int64
          description: 所属空间ID
          example: 1
        level:
          type: integer
          description: 文件夹层级
          example: 1
        isRoot:
          type: boolean
          description: 是否为根文件夹
          example: false
        fileCount:
          type: integer
          description: 包含文件数量
          example: 10
        sizeBytes:
          type: integer
          format: int64
          description: 文件夹总大小(字节)
          example: 10485760
        createdAt:
          type: string
          format: date-time
          description: 创建时间
          example: "2024-01-01T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
          example: "2024-01-01T12:00:00Z"

    # 创建文件夹请求
    CreateFolderRequest:
      type: object
      required:
        - name
        - spaceId
      properties:
        name:
          type: string
          description: 文件夹名称
          pattern: '^[^/\\:*?"<>|]+$'
          example: "新文件夹"
        spaceId:
          type: integer
          format: int64
          description: 所属空间ID
          example: 1
        parentFolderId:
          type: integer
          format: int64
          description: 父文件夹ID(可选)
          nullable: true
          example: 1

    # 文件夹树节点
    FolderTreeNode:
      type: object
      properties:
        folder:
          $ref: '#/components/schemas/FolderInfo'
        children:
          type: array
          description: 子文件夹
          items:
            $ref: '#/components/schemas/FolderTreeNode'
        totalSize:
          type: integer
          format: int64
          description: 总大小(包含子文件夹)
          example: 20971520
        totalFileCount:
          type: integer
          description: 总文件数(包含子文件夹)
          example: 25

    # ==================== 错误响应模型 ====================
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            success:
              type: boolean
              enum: [false]
            errors:
              type: array
              items:
                type: string

  # ==================== 通用参数 ====================
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码(从0开始)
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      description: 每页大小
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "请求参数验证失败"
            errors: ["用户名不能为空"]
            timestamp: 1642780800000
    Unauthorized:
      description: 未授权访问
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "访问令牌无效或已过期"
            timestamp: 1642780800000
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "权限不足，无法访问此资源"
            timestamp: 1642780800000
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "请求的资源不存在"
            timestamp: 1642780800000
...
paths:
...
  /api/folders:
    post:
      tags:
        - 文件夹管理
      summary: 创建文件夹
      description: |
        在指定空间和父文件夹下创建新文件夹。
        
        **注意事项**:
        - 文件夹名不能包含特殊字符
        - 同一父目录下不能有重名文件夹
        - 文件夹路径自动生成
      operationId: createFolder
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFolderRequest'
            example:
              name: "重要文档"
              spaceId: 1
              parentFolderId: null
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FolderInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: 文件夹名冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - 文件夹管理
      summary: 获取文件夹列表
      description: |
        获取指定空间下的文件夹列表，支持分页和排序。
        
        **查询功能**:
        - 按父文件夹过滤
        - 支持分页浏览
        - 多种排序方式
      operationId: getFolders
      security:
        - BearerAuth: []
      parameters:
        - name: spaceId
          in: query
          required: true
          description: 空间ID
          schema:
            type: integer
            format: int64
            example: 1
        - name: parentId
          in: query
          required: false
          description: 父文件夹ID(为空时获取根目录)
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: sort
          in: query
          required: false
          description: 排序字段
          schema:
            type: string
            enum: [name, createdAt, updatedAt, sizeBytes, fileCount]
            default: "name"
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          content:
                            type: array
                            items:
                              $ref: '#/components/schemas/FolderInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/folders/{folderId}:
    get:
      tags:
        - 文件夹管理
      summary: 获取文件夹详情
      description: |
        根据文件夹ID获取详细信息。
        
        **权限要求**: 需要对文件夹有读取权限
      operationId: getFolderDetail
      security:
        - BearerAuth: []
      parameters:
        - name: folderId
          in: path
          required: true
          description: 文件夹ID
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FolderInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - 文件夹管理
      summary: 更新文件夹信息
      description: |
        更新文件夹的名称等信息。
        
        **权限要求**: 需要对文件夹有写入权限
      operationId: updateFolder
      security:
        - BearerAuth: []
      parameters:
        - name: folderId
          in: path
          required: true
          description: 文件夹ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 新文件夹名称
                  example: "重要文档-已归档"
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FolderInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - 文件夹管理
      summary: 删除文件夹
      description: |
        删除指定文件夹。注意：删除文件夹会同时删除其中的所有文件和子文件夹。
        
        **权限要求**: 需要对文件夹有删除权限
        
        **危险操作**: 此操作不可逆，请谨慎使用
      operationId: deleteFolder
      security:
        - BearerAuth: []
      parameters:
        - name: folderId
          in: path
          required: true
          description: 文件夹ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 文件夹非空，无法删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/folders/tree:
    get:
      tags:
        - 文件夹管理
      summary: 获取文件夹树结构
      description: |
        获取指定空间的完整文件夹树结构。
        
        **功能特性**:
        - 层次化文件夹结构
        - 包含每个节点的统计信息
        - 支持深度限制
      operationId: getFolderTree
      security:
        - BearerAuth: []
      parameters:
        - name: spaceId
          in: query
          required: true
          description: 空间ID
          schema:
            type: integer
            format: int64
            example: 1
        - name: maxDepth
          in: query
          required: false
          description: 最大深度(-1表示无限制)
          schema:
            type: integer
            default: -1
            example: 3
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/FolderTreeNode'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== 系统配置API ====================
...